apps/subnet-calculator/frontend-html-static/pyproject.toml
[project]
name = "subnet-calculator-static-frontend"
version = "0.1.0"
description = "Static HTML frontend for IPv4 subnet calculator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "playwright>=1.55.0",
    "pytest-playwright>=0.7.1",
    "pytest-base-url>=2.1.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
